split routing out as a seperate entity
